python_script/requirement.txt
pyzk
mysql-connector
pymysql
python-dotenv